requirements.txt
# requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0